+++ CHANGELOG.md
@@ -3,11 +3,18 @@
 - [Jeeva] Added View and download button in the view report page.
 
 ## [1.0.0-Pre02] - 01-Apr-2026
-### Added
+### Fixed
 - [Jeeva] Fixed Word doc Download and Attachment for Mail recieve and response fields.
 
 ## [1.0.0-Pre03] - 13-Apr-2026
-### Added
+### Fixed
 - [Jeeva] Fixed Multi image upload and view and UI Update
 
 - [Jeeva] Fixed Image replace during edit and edit tab close.
+
+
+## [1.0.0-Pre04] - 15-Apr-2026
+### Added
+- [Jeeva] Login and signup page
+### Fixed
+fixed document download format